Move more vDSO functionality into vdso module
Move the Symbolizer::create_vdso_parser() method as a freestanding
function into the vdso module. This way, it can be repurposed more
easily.
